// src/types/types.ts
import { ResultOf } from "gql.tada";
import { GET_COUNTRIES, GET_COUNTRY_DETAILS } from "../graphql/queries";

export interface IWeatherData {
  coord: {
    lon: number;
    lat: number;
  };
  weather: Array<{
    id: number;
    main: string;
    description: string;
    icon: string;
  }>;
  base: string;
  main: {
    temp: number;
    feels_like: number;
    temp_min: number;
    temp_max: number;
    pressure: number;
    humidity: number;
    sea_level: number;
    grnd_level: number;
  };
  visibility: number;
  wind: {
    speed: number;
    deg: number;
    gust: number;
  };
  clouds: {
    all: number;
  };
  dt: number;
  sys: {
    country: string;
    sunrise: number;
    sunset: number;
  };
  timezone: number;
  id: number;
  name: string;
  cod: number;
}

export type TCountry = ResultOf<typeof GET_COUNTRIES>["countries"][0];
export type TCountryDetails = ResultOf<typeof GET_COUNTRY_DETAILS>["country"];